Added check for venv

+++ pythanos.py
@@ -11,25 +11,34 @@ class bcolors:
     BOLD = '\033[1m'
     UNDERLINE = '\033[4m'
     
+
+def inside_venv():
+    return sys.base_prefix != sys.prefix
+    
     
 def pythanos():
-    if not hasattr(sys, "real_prefix"):
+    if not inside_venv():
         # Check if you are inside a virtual environment
-        print(bcolors.WARNING, "Warning! You are outside a virtual environment. This will erase all system wide packages!", bcolor.ENDC)
+        print(bcolors.WARNING, "Warning! You are outside a virtual environment. This will erase all system wide packages!", bcolors.ENDC)
         sys.exit(1)
     else:
         curr_packages = subprocess.getoutput('pip freeze').split('\n')
         
-    print("Found {} installed packages in the current virtual environment".format(len(curr_packages))
+    print("Found {} installed packages in the current virtual environment".format(len(curr_packages)))
 
     # TODO - add an argument to get the packages before removing them
     # TODO - ask for confirmation
-    print ("Pythanos snaps...")
     
-    for package in curr_packages:
-        print("Removing {}".format(package))
-        command = 'pip uninstall {} -y'.format(package)
-        print(subprocess.getoutput(command))
+    c = input('Confirmation for Thanos to fuck this world up!')
+
+    if c == 'y':
+
+        print ("Pythanos snaps...")
+        
+        for package in curr_packages:
+            print("Removing {}".format(package))
+            command = 'pip uninstall {} -y'.format(package)
+            print(subprocess.getoutput(command))
       
 if __name__ == '__main__':
     pythanos()
